server: don't load config from disk on every request
In my first iteration I had every handler loading the config from disk
because I was too lazy to add a lock to a global singleton conf
instance. But now I figured it was time to fix that so I added a
goroutine mini-server that reads and writes the config in a threadsafe
and lockfree fashion.

Also, hooked a function to signal SIGUSR2 (`kill -USR2`) which reloads
the config from disk, if you made any changes. After sending the signal
you can `tail` your server logs and see the confirmation that it was
reloaded.